refactor : PawnMoveStrategy 를 Color 로 분리
