+++ src/main/java/chess/domain/move/FirstRowMoveStrategy.java
@@ -5,7 +5,6 @@ import chess.domain.piece.Piece;
 
 public abstract class FirstRowMoveStrategy extends MoveStrategy {
 
-    @Override
     protected boolean isMovableToTarget(Piece targetPiece, Color color) {
         if (!targetPiece.isBlank()) {
             return targetPiece.getColor() == color.oppositeColor();
